+++ microerror.go
@@ -9,7 +9,7 @@ import (
 
 // Cause is here only for backward compatibility purposes and should not be used.
 //
-// NOTE: Use Is instead.
+// NOTE: Use errors.Is/errors.As instead.
 func Cause(err error) error {
 	e := err
 	for e != nil {
@@ -20,11 +20,7 @@ func Cause(err error) error {
 	return err
 }
 
-func Is(err, target error) bool {
-	return errors.Is(err, target)
-}
-
-func Maskf(err *Error, f string, v ...interface{}) error {
+func Maskf(err Error, f string, v ...interface{}) error {
 	annotatedErr := annotatedError{
 		annotation: fmt.Sprintf(f, v...),
 		underlying: err,
@@ -56,8 +52,8 @@ func mask(err error) error {
 // New is here only for backward compatibility purposes and should not be used.
 //
 // NOTE: Use struct initialization literal for Error struct instead.
-func New(kind string) *Error {
-	return &Error{
+func New(kind string) Error {
+	return Error{
 		Kind: kind,
 	}
 }