+++ microerror_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func Test_Cause(t *testing.T) {
-	var testCauseMicroError = &Error{
+	var testCauseMicroError = Error{
 		Kind: "testCauseErrorB",
 	}
 	var testCauseErrorsNewError = errors.New("test cause error A")
@@ -115,7 +115,7 @@ func Test_Mask_Nil(t *testing.T) {
 	}
 }
 
-var testMicroErr = &Error{
+var testMicroErr = Error{
 	Desc: "test-desc",
 	Docs: "test-docs",
 	Kind: "testKind",